Copied Tuan's Useless Folder into the main folder; SEE DESCRIPTION
-First of all, changed the titlebar to say Greenchecks as well as the
title icon.
-Secondly, why did Tuan create a new directory in the repo with his
version of the program (without any kind of description of what it was
for that matter)? What was the point of creating this directory. At
least explain to me what you have changed and added.
-Third of all, dental benefits needs to be added to the main database
and the textbox needs to be added to Form1.cs
-Last, but not least, the check window needs to have a $ amount of these
types of benefits. That can come later. I am going to link this commit
message to an issue on Github so everyone can see.
-Tuan, please document your changes accordingly. Do not simply write
"New Update". That is insufficient to describe the actions you are
taking. Do not do this again, please.

Thank you.

-Chris